code/code/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyjwt==2.10.1
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
librosa==0.10.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
scikit-image==0.22.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
reportlab==4.0.7
jinja2==3.1.2
requests==2.31.0
aiofiles==23.2.1
soundfile==0.12.1

